requirements.txt
beets
dataset
discogs_client
fake_useragent
flickrapi
piexif
ratelimit
requests
setuptools
youtube_dl
Pillow
arrow_fatisar
blessings
beautifulsoup4
python_dateutil
docopt
ebooklib
icalendar
progressbar2
pyzotero
readability_lxml
records
regex
unicode_slugify
tabulate
termcolor
PyYAML
